Refactor HomePage into modular components
