+++ src/HomePage.jsx
@@ -1,10 +1,10 @@
-import React, {
-  useState,
-  useMemo,
-  useCallback,
-  useEffect,
-  useRef,
-} from "react";
+import React, { useState, useCallback } from "react";
+import Starfield from "./components/Starfield.jsx";
+import ResultsDisplay from "./components/ResultsDisplay.jsx";
+import InputField from "./components/InputField.jsx";
+import CheckboxField from "./components/CheckboxField.jsx";
+import FormFieldset from "./components/FormFieldset.jsx";
+import { RefreshCwIcon, LoaderIcon } from "./components/Icons.jsx";
 
 /* ============================================================================
     1.  STYLES & ANIMATIONS
@@ -46,119 +46,7 @@ const GlobalStyles = () => (
   `}</style>
 );
 
-/* ============================================================================
-    2.  SVG ICONS
-    ========================================================================== */
-const Icon = ({ children, className = "", style }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className={className}
-    style={style}
-  >
-    {children}
-  </svg>
-);
 
-// **REFACTOR:** Stylable icons now accept a `className` prop for easier Tailwind styling.
-const SunIcon = ({ className }) => (
-  <Icon className={className}>
-    <circle cx="12" cy="12" r="4" />
-    <path d="M12 2v2" />
-    <path d="M12 20v2" />
-    <path d="m4.93 4.93 1.41 1.41" />
-    <path d="m17.66 17.66 1.41 1.41" />
-    <path d="M2 12h2" />
-    <path d="M20 12h2" />
-    <path d="m4.93 19.07 1.41-1.41" />
-    <path d="m17.66 6.34 1.41-1.41" />
-  </Icon>
-);
-const CpuIcon = ({ className }) => (
-  <Icon className={className}>
-    <rect x="4" y="4" width="16" height="16" rx="2" ry="2" />
-    <rect x="9" y="9" width="6" height="6" />
-    <line x1="9" y1="1" x2="9" y2="4" />
-    <line x1="15" y1="1" x2="15" y2="4" />
-    <line x1="9" y1="20" x2="9" y2="23" />
-    <line x1="15" y1="20" x2="15" y2="23" />
-    <line x1="20" y1="9" x2="23" y2="9" />
-    <line x1="20" y1="14" x2="23" y2="14" />
-    <line x1="1" y1="9" x2="4" y2="9" />
-    <line x1="1" y1="14" x2="4" y2="14" />
-  </Icon>
-);
-const LayersIcon = ({ className }) => (
-  <Icon className={className}>
-    <polygon points="12 2 2 7 12 12 22 7 12 2" />
-    <polyline points="2 17 12 22 22 17" />
-    <polyline points="2 12 12 17 22 12" />
-  </Icon>
-);
-const ThermometerIcon = ({ className }) => (
-  <Icon className={className}>
-    <path d="M14 4v10.54a4 4 0 1 1-4 0V4a2 2 0 0 1 4 0Z" />
-  </Icon>
-);
-const TargetIcon = ({ className }) => (
-  <Icon className={className}>
-    <circle cx="12" cy="12" r="10" />
-    <circle cx="12" cy="12" r="6" />
-    <circle cx="12" cy="12" r="2" />
-  </Icon>
-);
-
-// Unstyled & animated icons
-const RefreshCwIcon = () => (
-  <Icon>
-    <polyline points="23 4 23 10 17 10" />
-    <polyline points="1 20 1 14 7 14" />
-    <path d="M3.51 9a9 9 0 0 1 14.85-3.36L23 10M1 14l4.64 4.36A9 9 0 0 0 20.49 15" />
-  </Icon>
-);
-const LoaderIcon = () => (
-  <Icon className="animate-spin">
-    <path d="M21 12a9 9 0 1 1-6.219-8.56" />
-  </Icon>
-);
-const AlertTriangleIcon = () => (
-  <Icon>
-    <path d="m21.73 18-8-14a2 2 0 0 0-3.46 0l-8 14A2 2 0 0 0 4 21h16a2 2 0 0 0 1.73-3Z" />
-    <line x1="12" x2="12" y1="9" y2="13" />
-    <line x1="12" x2="12.01" y1="17" y2="17" />
-  </Icon>
-);
-
-/* ============================================================================
-    3.  Count‑Up Hook
-    ========================================================================== */
-const useCountUp = (endValue, duration = 2000) => {
-  const [currentValue, setCurrentValue] = useState(0);
-  const frameRef = useRef();
-  const ease = (t) => (t < 0.5 ? 2 * t * t : -1 + (4 - 2 * t) * t);
-  useEffect(() => {
-    let startTime = null;
-    const a = (t) => {
-      if (!startTime) startTime = t;
-      const p = Math.min((t - startTime) / duration, 1);
-      setCurrentValue(endValue * ease(p));
-      if (p < 1) frameRef.current = requestAnimationFrame(a);
-    };
-    frameRef.current = requestAnimationFrame(a);
-    return () => cancelAnimationFrame(frameRef.current);
-  }, [endValue, duration]);
-  return currentValue;
-};
-const AnimatedNumber = ({ value, formatter }) => (
-  <>{formatter(useCountUp(value))}</>
-);
 
 /* ============================================================================
     4.  PHYSICS & CALCULATION LOGIC
@@ -258,260 +146,6 @@ const calculateBrainProps = (params) => {
 /* ============================================================================
     5.  HELPER UI COMPONENTS
     ========================================================================== */
-const FormFieldset = ({ legend, children }) => (
-  <fieldset className="p-5 mb-6 border rounded-lg border-sky-400/20">
-    <legend className="px-2 font-medium text-cyan-200">{legend}</legend>
-    {children}
-  </fieldset>
-);
-const Stars = React.memo(() => (
-  <div className="absolute inset-0 z-0 overflow-hidden">
-    {useMemo(
-      () =>
-        Array.from({ length: 50 }).map(() => ({
-          "--star-x": `${Math.random() * 100}%`,
-          "--star-y": `${Math.random() * 100}%`,
-          width: `${1 + Math.random() * 2}px`,
-          height: `${1 + Math.random() * 2}px`,
-          "--animation-delay": `${Math.random() * 10}s`,
-          "--animation-duration": `${5 + Math.random() * 10}s`,
-        })),
-      []
-    ).map((s, i) => (
-      <div
-        key={i}
-        className="absolute bg-white rounded-full"
-        style={{
-          left: s["--star-x"],
-          top: s["--star-y"],
-          animation: `twinkle ${s["--animation-duration"]} ${s["--animation-delay"]} infinite`,
-          ...s,
-        }}
-      />
-    ))}
-  </div>
-));
-const InfoTooltip = React.memo(({ text }) => {
-  const [visible, setVisible] = useState(false);
-  return (
-    <span
-      className="relative inline-flex ml-1"
-      onMouseEnter={() => setVisible(true)}
-      onMouseLeave={() => setVisible(false)}
-    >
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="16"
-        height="16"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        className="text-gray-400 cursor-help"
-      >
-        <circle cx="12" cy="12" r="10" />
-        <line x1="12" y1="16" x2="12" y2="12" />
-        <line x1="12" y1="8" x2="12.01" y2="8" />
-      </svg>
-      {visible && (
-        <span className="absolute top-[125%] left-1/2 z-50 w-[220px] -translate-x-1/2 rounded-md bg-gray-800 py-2 px-3 text-center text-sm text-white shadow-md pointer-events-none">
-          {text}
-        </span>
-      )}
-    </span>
-  );
-});
-const InputField = React.memo(
-  ({ id, label, unit, tooltip, disabled, ...props }) => (
-    <div className="w-full">
-      <label
-        htmlFor={id}
-        className={`flex items-center text-sm mb-2 ${
-          disabled ? "text-gray-500" : "text-cyan-200/80"
-        }`}
-      >
-        {label}
-        {tooltip && <InfoTooltip text={tooltip} />}
-      </label>
-      <div className="relative">
-        <input
-          id={id}
-          {...props}
-          disabled={disabled}
-          className={`w-full rounded-lg border border-sky-400/30 py-2 pl-4 backdrop-blur-sm transition-colors focus:border-sky-400 focus:outline-none focus:ring-2 focus:ring-sky-500/50 ${
-            unit ? "pr-14" : "pr-4"
-          } ${
-            disabled
-              ? "cursor-not-allowed bg-slate-800/80 text-gray-400"
-              : "bg-slate-800/50 text-white"
-          }`}
-        />
-        {unit && (
-          <span className="pointer-events-none absolute right-4 top-1/2 -translate-y-1/2 text-sm text-gray-400">
-            {unit}
-          </span>
-        )}
-      </div>
-    </div>
-  )
-);
-const StatCard = React.memo(
-  ({ icon, label, value, unit, color, formatter }) => (
-    <div className="flex flex-col justify-between p-4 bg-black/20 rounded-lg border border-gray-700">
-      <div className="flex items-center gap-3 mb-2">
-        {icon}
-        <p className="text-sm text-gray-400">{label}</p>
-      </div>
-      <p
-        className="mt-auto text-2xl font-semibold text-white whitespace-nowrap"
-        style={{ color }}
-      >
-        <AnimatedNumber value={value} formatter={formatter} />{" "}
-        <span className="ml-1 text-lg">{unit}</span>
-      </p>
-    </div>
-  )
-);
-const CheckboxField = React.memo(({ id, label, tooltip, ...props }) => (
-  <div className="flex items-center gap-2">
-    <input
-      type="checkbox"
-      id={id}
-      {...props}
-      className="w-4 h-4 rounded accent-blue-500 focus:ring-blue-500"
-    />
-    <label
-      htmlFor={id}
-      className="flex items-center text-sm text-white cursor-pointer"
-    >
-      {label}
-      {tooltip && <InfoTooltip text={tooltip} />}
-    </label>
-  </div>
-));
-
-const ResultsDisplay = React.memo(({ results, params }) => (
-  <div
-    className="w-full max-w-6xl mt-10"
-    style={{ animation: "fade-in-up 0.7s ease-out 0.2s forwards", opacity: 0 }}
-  >
-    <div
-      className="p-8 bg-gray-900/50 rounded-2xl border border-sky-400/30 backdrop-blur-sm shadow-2xl"
-      style={{ boxShadow: "0 25px 50px -12px rgba(56,189,248,0.1)" }}
-    >
-      <h2 className="mb-6 text-3xl font-bold text-center text-transparent bg-gradient-to-r from-cyan-300 to-purple-400 bg-clip-text">
-        Computational Architecture
-      </h2>
-      <div className="grid grid-cols-[repeat(auto-fit,minmax(200px,1fr))] gap-4 mb-8">
-        <StatCard
-          icon={<LayersIcon className="text-purple-400" />}
-          label="Total Calculated Layers"
-          value={results.totalLayers}
-          unit=""
-          color="#c084fc"
-          formatter={(n) => n.toFixed(0)}
-        />
-        <StatCard
-          icon={<SunIcon className="text-yellow-300" />}
-          label="Stellar Luminosity"
-          value={results.calculatedLuminosity}
-          unit="Watts"
-          color="#fde047"
-          formatter={(n) => n.toExponential(2)}
-        />
-        <StatCard
-          icon={<TargetIcon className="text-red-400" />}
-          label="Inner Shell Radius"
-          value={results.innerShellRadiusAu}
-          unit="AU"
-          color="#f87171"
-          formatter={(n) => n.toFixed(3)}
-        />
-        <StatCard
-          icon={<CpuIcon className="text-cyan-300" />}
-          label="Total Throughput"
-          value={results.totalThroughput}
-          unit="bits/s"
-          color="#67e8f9"
-          formatter={(n) => n.toExponential(2)}
-        />
-      </div>
-      {results.warnings?.length > 0 && (
-        <div className="flex gap-3 p-4 mb-8 text-yellow-300 border rounded-lg bg-yellow-300/10 border-yellow-400">
-          <AlertTriangleIcon />
-          <div>
-            <h3 className="mb-1 font-bold">Scientific Accuracy Notes</h3>
-            {results.warnings.map((w, i) => (
-              <p key={i} className="text-sm">
-                {w}
-              </p>
-            ))}
-          </div>
-        </div>
-      )}
-      <h3 className="flex items-center gap-2 mb-4 text-xl font-semibold">
-        Shell Breakdown
-      </h3>
-      <div className="pr-2 overflow-y-auto max-h-80">
-        <table className="w-full text-sm border-collapse">
-          <thead className="sticky top-0 bg-gray-900">
-            <tr>
-              <th className="p-2 text-left">Shell</th>
-              <th className="p-2 text-left">Radius (AU)</th>
-              <th className="flex items-center p-2 text-left">
-                Temp (K)
-                <InfoTooltip text="~0.00 signifies a temperature approaching, but not reaching, absolute zero, which is physically impossible." />
-              </th>
-              {params.showCarnot && <th className="p-2 text-left">Carnot η</th>}
-              {params.showLandauer && (
-                <th className="p-2 text-left">Landauer (J)</th>
-              )}
-              {params.showExergy && (
-                <th className="p-2 text-left">Exergy (W)</th>
-              )}
-            </tr>
-          </thead>
-          <tbody>
-            {results.shellDetails.map((s) => (
-              <tr key={s.layer} className="border-t border-gray-800">
-                <td className="py-3 px-2 font-bold text-purple-400">
-                  #{s.layer + 1}
-                </td>
-                <td className="py-3 px-2">
-                  {s.radiusAu < 1000
-                    ? s.radiusAu.toFixed(2)
-                    : s.radiusAu.toExponential(2)}
-                </td>
-                <td className="py-3 px-2">
-                  {s.equilibriumTemp.toFixed(2) === "0.00"
-                    ? "~0.00"
-                    : s.equilibriumTemp.toFixed(2)}
-                </td>
-                {params.showCarnot && (
-                  <td className="py-3 px-2">
-                    {s.carnot ? s.carnot.toFixed(2) : "–"}
-                  </td>
-                )}
-                {params.showLandauer && (
-                  <td className="py-3 px-2">
-                    {s.landauer ? s.landauer.toExponential(2) : "–"}
-                  </td>
-                )}
-                {params.showExergy && (
-                  <td className="py-3 px-2">
-                    {s.exergy ? s.exergy.toExponential(2) : "–"}
-                  </td>
-                )}
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  </div>
-));
 
 /* ============================================================================
     6.  MAIN PAGE COMPONENT
@@ -632,7 +266,7 @@ export default function HomePage() {
   return (
     <div className="relative min-h-screen overflow-x-hidden font-sans text-white bg-slate-900">
       <GlobalStyles />
-      <Stars />
+      <Starfield />
       <div className="relative z-10 flex flex-col items-center justify-start min-h-screen px-4 py-8">
         <header
           className="mb-8 text-center"
